Propagate the writer error to the reader
Use CloseWithError in the writer end of the pipe to inform the reader
end that there was an error in the process.
Use a specific type of error, sent by the writer and checked by the
reader, to make sure that only one error post is sent to the user.