MM-25396: emit RECEIVED_WEBAPP_PLUGIN
As part of https://mattermost.atlassian.net/browse/MM-17087, the webapp no longer emits `RECEIVED_WEBAPP_PLUGIN`. This makes it impossible to use the Redux store alone to detect the presence of plugins installed /after/ a page load. (The webapp still tracks all such plugins via `RECEIVED_WEBAPP_PLUGINS` on page load.)

Emit the event ourselves to better interop with the incident response plugin while we wait for the webapp to fix this.

Fixes: https://mattermost.atlassian.net/browse/MM-25396